+++ demo/nextjs/lib/auth.ts
@@ -8,6 +8,7 @@ import {
 	oneTap,
 	oAuthProxy,
 	openAPI,
+	oidcProvider,
 } from "better-auth/plugins";
 import { reactInvitationEmail } from "./email/invitation";
 import { LibsqlDialect } from "@libsql/kysely-libsql";
@@ -41,22 +42,6 @@ export const auth = betterAuth({
 	database: {
 		dialect,
 		type: "sqlite",
-	},
-	databaseHooks: {
-		user: {
-			update: {
-				async before(user) {
-					if (user.emailVerified) {
-						return {
-							data: {
-								...user,
-								emailVerifiedAt: new Date().toISOString(),
-							},
-						};
-					}
-				},
-			},
-		},
 	},
 	emailVerification: {
 		async sendVerificationEmail({ user, url }) {
@@ -161,5 +146,8 @@ export const auth = betterAuth({
 		oneTap(),
 		oAuthProxy(),
 		nextCookies(),
+		oidcProvider({
+			loginPage: "/sign-in",
+		}),
 	],
 });
