--- packages/better-auth/src/plugins/oidc/index.ts
@@ -1,710 +1,0 @@
-import { base64url, SignJWT } from "jose";
-import { z } from "zod";
-import { APIError, createAuthEndpoint, sessionMiddleware } from "../../api";
-import type { BetterAuthPlugin, GenericEndpointContext } from "../../types";
-import { generateRandomString } from "../../crypto";
-import { schema } from "./schema";
-import type {
-	Client,
-	CodeVerificationValue,
-	OAuthAccessToken,
-	OIDCMetadata,
-	OIDCOptions,
-} from "./types";
-import { authorize } from "./authorize";
-import { parseSetCookieHeader } from "../../cookies";
-import { sha256 } from "oslo/crypto";
-import type { Endpoint } from "better-call";
-
-const getMetadata = (
-	ctx: GenericEndpointContext,
-	options?: OIDCOptions,
-): OIDCMetadata => {
-	const issuer = ctx.context.options.baseURL as string;
-	const baseURL = ctx.context.baseURL;
-	return {
-		issuer,
-		authorization_endpoint: `${baseURL}/oauth2/authorize`,
-		token_endpoint: `${baseURL}/oauth2/token`,
-		userInfo_endpoint: `${baseURL}/oauth2/userinfo`,
-		jwks_uri: `${baseURL}/jwks`,
-		registration_endpoint: `${baseURL}/oauth2/register`,
-		scopes_supported: ["openid", "profile", "email", "offline_access"],
-		response_types_supported: ["code"],
-		response_modes_supported: ["query"],
-		grant_types_supported: ["authorization_code"],
-		acr_values_supported: [
-			"urn:mace:incommon:iap:silver",
-			"urn:mace:incommon:iap:bronze",
-		],
-		subject_types_supported: ["public"],
-		id_token_signing_alg_values_supported: ["RS256", "none"],
-		token_endpoint_auth_methods_supported: [
-			"client_secret_basic",
-			"client_secret_post",
-		],
-		claims_supported: [
-			"sub",
-			"iss",
-			"aud",
-			"exp",
-			"nbf",
-			"iat",
-			"jti",
-			"email",
-			"email_verified",
-			"name",
-		],
-		...options?.metadata,
-	};
-};
-
-/**
- * OpenID Connect (OIDC) plugin for Better Auth. This plugin implements the
- * authorization code flow and the token exchange flow. It also implements the
- * userinfo endpoint.
- *
- * @param options - The options for the OIDC plugin.
- * @returns A Better Auth plugin.
- */
-export const oidcProvider = (options: OIDCOptions) => {
-	const modelName = {
-		oauthClient: "oauthClient",
-		oauthAccessToken: "oauthAccessToken",
-	};
-
-	const opts = {
-		codeExpiresIn: 600,
-		defaultScope: "openid",
-		accessTokenExpiresIn: 3600,
-		refreshTokenExpiresIn: 604800,
-		...options,
-		scopes: [
-			"openid",
-			"profile",
-			"email",
-			"offline_access",
-			...(options?.scopes || []),
-		],
-	};
-
-	return {
-		id: "oidc",
-		hooks: {
-			after: [
-				{
-					matcher() {
-						return true;
-					},
-					handler: async (ctx) => {
-						const cookie = await ctx.getSignedCookie(
-							"oidc_login_prompt",
-							ctx.context.secret,
-						);
-						const cookieName = ctx.context.authCookies.sessionToken.name;
-						const parsedSetCookieHeader = parseSetCookieHeader(
-							ctx.responseHeader.get("set-cookie") || "",
-						);
-						const hasSessionToken = parsedSetCookieHeader.has(cookieName);
-						if (!cookie || !hasSessionToken) {
-							return;
-						}
-						ctx.setCookie("oidc_login_prompt", "", {
-							maxAge: 0,
-						});
-						const sessionCookie = parsedSetCookieHeader.get(cookieName)?.value;
-						const sessionToken = sessionCookie?.split(".")[0];
-						if (!sessionToken) {
-							return;
-						}
-						const session =
-							await ctx.context.internalAdapter.findSession(sessionToken);
-						if (!session) {
-							return;
-						}
-						ctx.query = JSON.parse(cookie);
-						ctx.query.prompt = "consent";
-						ctx.context.session = session;
-						const response = await authorize(ctx, opts);
-						return response;
-					},
-				},
-			],
-		},
-		endpoints: {
-			getOpenIdConfig: createAuthEndpoint(
-				"/.well-known/openid-configuration",
-				{
-					method: "GET",
-				},
-				async (ctx) => {
-					const metadata = getMetadata(ctx, options);
-					return metadata;
-				},
-			),
-			oAuth2authorize: createAuthEndpoint(
-				"/oauth2/authorize",
-				{
-					method: "GET",
-					query: z.record(z.string(), z.any()),
-				},
-				async (ctx) => {
-					return authorize(ctx, opts);
-				},
-			),
-			oAuthConsent: createAuthEndpoint(
-				"/oauth2/consent",
-				{
-					method: "POST",
-					body: z.object({
-						accept: z.boolean(),
-					}),
-					use: [sessionMiddleware],
-				},
-				async (ctx) => {
-					const storedCode = await ctx.getSignedCookie(
-						"oidc_consent_prompt",
-						ctx.context.secret,
-					);
-					if (!storedCode) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "No consent prompt found",
-							error: "invalid_grant",
-						});
-					}
-					const verification =
-						await ctx.context.internalAdapter.findVerificationValue(storedCode);
-					if (!verification) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "Invalid code",
-							error: "invalid_grant",
-						});
-					}
-					if (verification.expiresAt < new Date()) {
-						await ctx.context.internalAdapter.deleteVerificationValue(
-							verification.id,
-						);
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "Code expired",
-							error: "invalid_grant",
-						});
-					}
-					const value = JSON.parse(verification.value) as CodeVerificationValue;
-					if (!value.requireConsent || !value.state) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "Consent not required",
-							error: "invalid_grant",
-						});
-					}
-
-					if (!ctx.body.accept) {
-						await ctx.context.internalAdapter.deleteVerificationValue(
-							verification.id,
-						);
-						return ctx.json({
-							redirectURI: `${value.redirectURI}?error=access_denied&error_description=User denied access`,
-						});
-					}
-					const code = generateRandomString(32, "a-z", "A-Z", "0-9");
-					const codeExpiresInMs = opts.codeExpiresIn * 1000;
-					const expiresAt = new Date(Date.now() + codeExpiresInMs);
-					await ctx.context.internalAdapter.updateVerificationValue(
-						verification.id,
-						{
-							value: JSON.stringify({
-								...value,
-								requireConsent: false,
-							}),
-							identifier: code,
-							expiresAt,
-						},
-					);
-					const redirectURI = new URL(value.redirectURI);
-					redirectURI.searchParams.set("code", code);
-					redirectURI.searchParams.set("state", value.state);
-					return ctx.json({
-						redirectURI: redirectURI.toString(),
-					});
-				},
-			),
-			oAuth2token: createAuthEndpoint(
-				"/oauth2/token",
-				{
-					method: "POST",
-					body: z.any(),
-					metadata: {
-						isAction: false,
-					},
-				},
-				async (ctx) => {
-					let { body } = ctx;
-					if (!body) {
-						throw new APIError("BAD_REQUEST", {
-							error_description: "request body not found",
-							error: "invalid_request",
-						});
-					}
-					if (body instanceof FormData) {
-						body = Object.fromEntries(body.entries());
-					}
-					if (!(body instanceof Object)) {
-						throw new APIError("BAD_REQUEST", {
-							error_description: "request body is not an object",
-							error: "invalid_request",
-						});
-					}
-					const {
-						client_id,
-						client_secret,
-						grant_type,
-						code,
-						redirect_uri,
-						refresh_token,
-						code_verifier,
-					} = body;
-					if (grant_type === "refresh_token") {
-						if (!refresh_token) {
-							throw new APIError("BAD_REQUEST", {
-								error_description: "refresh_token is required",
-								error: "invalid_request",
-							});
-						}
-						const token = await ctx.context.adapter.findOne<OAuthAccessToken>({
-							model: modelName.oauthAccessToken,
-							where: [
-								{
-									field: "refreshToken",
-									value: refresh_token.toString(),
-								},
-							],
-						});
-						if (!token) {
-							throw new APIError("UNAUTHORIZED", {
-								error_description: "invalid refresh token",
-								error: "invalid_grant",
-							});
-						}
-						if (token.clientId !== client_id?.toString()) {
-							throw new APIError("UNAUTHORIZED", {
-								error_description: "invalid client_id",
-								error: "invalid_client",
-							});
-						}
-						if (token.refreshTokenExpiresAt < new Date()) {
-							throw new APIError("UNAUTHORIZED", {
-								error_description: "refresh token expired",
-								error: "invalid_grant",
-							});
-						}
-						const accessToken = generateRandomString(32, "a-z", "A-Z");
-						const newRefreshToken = generateRandomString(32, "a-z", "A-Z");
-						const accessTokenExpiresAt = new Date(
-							Date.now() + opts.accessTokenExpiresIn * 1000,
-						);
-						const refreshTokenExpiresAt = new Date(
-							Date.now() + opts.refreshTokenExpiresIn * 1000,
-						);
-						await ctx.context.adapter.create({
-							model: modelName.oauthAccessToken,
-							data: {
-								accessToken,
-								refreshToken: newRefreshToken,
-								accessTokenExpiresAt,
-								refreshTokenExpiresAt,
-								clientId: client_id.toString(),
-								userId: token.userId,
-								scopes: token.scopes,
-							},
-						});
-						return ctx.json({
-							access_token: accessToken,
-							token_type: "bearer",
-							expires_in: opts.accessTokenExpiresIn,
-							refresh_token: newRefreshToken,
-							scope: token.scopes,
-						});
-					}
-
-					if (!code) {
-						throw new APIError("BAD_REQUEST", {
-							error_description: "code is required",
-							error: "invalid_request",
-						});
-					}
-
-					if (options.requirePKCE && !code_verifier) {
-						throw new APIError("BAD_REQUEST", {
-							error_description: "code verifier is missing",
-							error: "invalid_request",
-						});
-					}
-
-					/**
-					 * We need to check if the code is valid before we can proceed
-					 * with the rest of the request.
-					 */
-					const verificationValue =
-						await ctx.context.internalAdapter.findVerificationValue(
-							code.toString(),
-						);
-					if (!verificationValue) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "invalid code",
-							error: "invalid_grant",
-						});
-					}
-					if (verificationValue.expiresAt < new Date()) {
-						await ctx.context.internalAdapter.deleteVerificationValue(
-							verificationValue.id,
-						);
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "code expired",
-							error: "invalid_grant",
-						});
-					}
-
-					await ctx.context.internalAdapter.deleteVerificationValue(
-						verificationValue.id,
-					);
-					if (!client_id || !client_secret) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "client_id and client_secret are required",
-							error: "invalid_client",
-						});
-					}
-					if (!grant_type) {
-						throw new APIError("BAD_REQUEST", {
-							error_description: "grant_type is required",
-							error: "invalid_request",
-						});
-					}
-					if (grant_type !== "authorization_code") {
-						throw new APIError("BAD_REQUEST", {
-							error_description: "grant_type must be 'authorization_code'",
-							error: "unsupported_grant_type",
-						});
-					}
-
-					if (!redirect_uri) {
-						throw new APIError("BAD_REQUEST", {
-							error_description: "redirect_uri is required",
-							error: "invalid_request",
-						});
-					}
-
-					const client = await ctx.context.adapter
-						.findOne<Record<string, any>>({
-							model: modelName.oauthClient,
-							where: [{ field: "clientId", value: client_id.toString() }],
-						})
-						.then((res) => {
-							if (!res) {
-								return null;
-							}
-							return {
-								...res,
-								redirectURLs: res.redirectURLs.split(","),
-								metadata: res.metadata ? JSON.parse(res.metadata) : {},
-							} as Client;
-						});
-					if (!client) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "invalid client_id",
-							error: "invalid_client",
-						});
-					}
-					if (client.disabled) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "client is disabled",
-							error: "invalid_client",
-						});
-					}
-					const isValidSecret =
-						client.clientSecret === client_secret.toString();
-					if (!isValidSecret) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "invalid client_secret",
-							error: "invalid_client",
-						});
-					}
-					const value = JSON.parse(
-						verificationValue.value,
-					) as CodeVerificationValue;
-					if (value.clientId !== client_id.toString()) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "invalid client_id",
-							error: "invalid_client",
-						});
-					}
-					if (value.redirectURI !== redirect_uri.toString()) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "invalid redirect_uri",
-							error: "invalid_client",
-						});
-					}
-					if (value.codeChallenge && !code_verifier) {
-						throw new APIError("BAD_REQUEST", {
-							error_description: "code verifier is missing",
-							error: "invalid_request",
-						});
-					}
-
-					const challenge =
-						value.codeChallengeMethod === "plain"
-							? code_verifier
-							: Buffer.from(
-									await sha256(new TextEncoder().encode(code_verifier || "")),
-								).toString("base64url");
-
-					if (challenge !== value.codeChallenge) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "code verification failed",
-							error: "invalid_request",
-						});
-					}
-
-					const requestedScopes = value.scope;
-					await ctx.context.internalAdapter.deleteVerificationValue(
-						code.toString(),
-					);
-					const accessToken = generateRandomString(32, "a-z", "A-Z");
-					const refreshToken = generateRandomString(32, "A-Z", "a-z");
-					const accessTokenExpiresAt = new Date(
-						Date.now() + opts.accessTokenExpiresIn * 1000,
-					);
-					const refreshTokenExpiresAt = new Date(
-						Date.now() + opts.refreshTokenExpiresIn * 1000,
-					);
-					await ctx.context.adapter.create({
-						model: modelName.oauthAccessToken,
-						data: {
-							accessToken,
-							refreshToken,
-							accessTokenExpiresAt,
-							refreshTokenExpiresAt,
-							clientId: client_id.toString(),
-							userId: value.userId,
-							scopes: requestedScopes.join(" "),
-						},
-					});
-					const user = await ctx.context.internalAdapter.findUserById(
-						value.userId,
-					);
-					if (!user) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "user not found",
-							error: "invalid_grant",
-						});
-					}
-					let secretKey = {
-						alg: "HS256",
-						key: await crypto.subtle.generateKey(
-							{
-								name: "HMAC",
-								hash: "SHA-256",
-							},
-							true,
-							["sign", "verify"],
-						),
-					};
-					const profile = {
-						given_name: user.name.split(" ")[0],
-						family_name: user.name.split(" ")[1],
-						name: user.name,
-						profile: user.image,
-						updated_at: user.updatedAt.toISOString(),
-					};
-					const email = {
-						email: user.email,
-						email_verified: user.emailVerified,
-					};
-					const userClaims = {
-						...(requestedScopes.includes("profile") ? profile : {}),
-						...(requestedScopes.includes("email") ? email : {}),
-					};
-
-					const idToken = await new SignJWT({
-						sub: user.id,
-						aud: client_id.toString(),
-						iat: Date.now(),
-						auth_time: ctx.context.session?.session.createdAt.getTime(),
-						nonce: body.nonce,
-						acr: "urn:mace:incommon:iap:silver", // default to silver - ⚠︎ this should be configurable and should be validated against the client's metadata
-						...userClaims,
-					})
-						.setProtectedHeader({ alg: secretKey.alg })
-						.setIssuedAt()
-						.setExpirationTime(
-							Math.floor(Date.now() / 1000) + opts.accessTokenExpiresIn,
-						)
-						.sign(secretKey.key);
-
-					return ctx.json(
-						{
-							access_token: accessToken,
-							token_type: "Bearer",
-							expires_in: opts.accessTokenExpiresIn,
-							refresh_token: requestedScopes.includes("offline_access")
-								? refreshToken
-								: undefined,
-							scope: requestedScopes.join(" "),
-							id_token: requestedScopes.includes("openid")
-								? idToken
-								: undefined,
-						},
-						{
-							headers: {
-								"Cache-Control": "no-store",
-								Pragma: "no-cache",
-							},
-						},
-					);
-				},
-			),
-			oAuth2userInfo: createAuthEndpoint(
-				"/oauth2/userinfo",
-				{
-					method: "GET",
-					metadata: {
-						isAction: false,
-					},
-				},
-				async (ctx) => {
-					if (!ctx.request) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "request not found",
-							error: "invalid_request",
-						});
-					}
-					const authorization = ctx.request.headers.get("authorization");
-					if (!authorization) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "authorization header not found",
-							error: "invalid_request",
-						});
-					}
-					const token = authorization.replace("Bearer ", "");
-					const accessToken =
-						await ctx.context.adapter.findOne<OAuthAccessToken>({
-							model: modelName.oauthAccessToken,
-							where: [
-								{
-									field: "accessToken",
-									value: token,
-								},
-							],
-						});
-					if (!accessToken) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "invalid access token",
-							error: "invalid_token",
-						});
-					}
-					if (accessToken.accessTokenExpiresAt < new Date()) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "The Access Token expired",
-							error: "invalid_token",
-						});
-					}
-
-					const user = await ctx.context.internalAdapter.findUserById(
-						accessToken.userId,
-					);
-					if (!user) {
-						throw new APIError("UNAUTHORIZED", {
-							error_description: "user not found",
-							error: "invalid_token",
-						});
-					}
-					const requestedScopes = accessToken.scopes.split(" ");
-					const userClaims = {
-						email: requestedScopes.includes("email") ? user.email : undefined,
-						name: requestedScopes.includes("profile") ? user.name : undefined,
-						picture: requestedScopes.includes("profile")
-							? user.image
-							: undefined,
-						given_name: requestedScopes.includes("profile")
-							? user.name.split(" ")[0]
-							: undefined,
-						family_name: requestedScopes.includes("profile")
-							? user.name.split(" ")[1]
-							: undefined,
-						email_verified: requestedScopes.includes("email")
-							? user.emailVerified
-							: undefined,
-					};
-					return ctx.json(userClaims);
-				},
-			),
-			createOAuthApplication: createAuthEndpoint(
-				"/oauth2/create-application",
-				{
-					method: "POST",
-					body: z.object({
-						name: z.string(),
-						icon: z.string().optional(),
-						metadata: z.record(z.any()).optional(),
-						redirectURLs: z.array(z.string()),
-					}),
-					use: [sessionMiddleware],
-				},
-				async (ctx) => {
-					const body = ctx.body;
-					const clientId =
-						options.generateClientId?.() ||
-						generateRandomString(32, "a-z", "A-Z");
-					const clientSecret =
-						options.generateClientSecret?.() ||
-						generateRandomString(32, "a-z", "A-Z");
-					const client = await ctx.context.adapter.create<Record<string, any>>({
-						model: modelName.oauthClient,
-						data: {
-							name: body.name,
-							icon: body.icon,
-							metadata: body.metadata ? JSON.stringify(body.metadata) : null,
-							clientId: clientId,
-							clientSecret: clientSecret,
-							redirectURLs: body.redirectURLs.join(","),
-							type: "web",
-							authenticationScheme: "client_secret",
-							disabled: false,
-							userId: ctx.context.session.user.id,
-						},
-					});
-					return ctx.json({
-						...client,
-						redirectURLs: client.redirectURLs.split(","),
-						metadata: client.metadata ? JSON.parse(client.metadata) : null,
-					} as Client);
-				},
-			),
-			getOAuthClient: createAuthEndpoint(
-				"/oauth2/client/:id",
-				{
-					method: "GET",
-					use: [sessionMiddleware],
-				},
-				async (ctx) => {
-					const client = await ctx.context.adapter.findOne<Record<string, any>>(
-						{
-							model: modelName.oauthClient,
-							where: [{ field: "clientId", value: ctx.params.id }],
-						},
-					);
-					if (!client) {
-						throw new APIError("NOT_FOUND", {
-							error_description: "client not found",
-							error: "not_found",
-						});
-					}
-					return ctx.json({
-						clientId: client.clientId as string,
-						name: client.name as string,
-						icon: client.icon as string,
-					});
-				},
-			),
-		},
-		schema,
-	} satisfies BetterAuthPlugin;
-};
